MNT: centralize time munging (again)
