+++ databroker/core.py
@@ -1,13 +1,11 @@
 from __future__ import print_function
 import six  # noqa
 from collections import defaultdict, deque
-from itertools import chain
-import pandas as pd
+from datetime import datetime
+import pytz
 import doct
 from pims import FramesSequence, Frame
 import logging
-import boltons.cacheutils
-import re
 import attr
 
 # Toolz and CyToolz have identical APIs -- same test suite, docstrings.
@@ -401,3 +399,100 @@ class DocBuffer:
                 yield doc
             else:
                 self.__stash_values(name, doc)
+
+
+def format_time(search_dict, tz):
+    """Helper function to format the time arguments in a search dict
+
+    Expects 'start_time' and 'stop_time'
+
+    ..warning: Does in-place mutation of the search_dict
+    """
+    time_dict = {}
+    start_time = search_dict.pop('start_time', None)
+    stop_time = search_dict.pop('stop_time', None)
+    if start_time:
+        time_dict['$gte'] = _normalize_human_friendly_time(start_time, tz)
+    if stop_time:
+        time_dict['$lte'] = _normalize_human_friendly_time(stop_time, tz)
+    if time_dict:
+        search_dict['time'] = time_dict
+
+
+# human friendly timestamp formats we'll parse
+_TS_FORMATS = [
+    '%Y-%m-%d %H:%M:%S',
+    '%Y-%m-%d %H:%M',  # these 2 are not as originally doc'd,
+    '%Y-%m-%d %H',     # but match previous pandas behavior
+    '%Y-%m-%d',
+    '%Y-%m',
+    '%Y']
+
+# build a tab indented, '-' bulleted list of supported formats
+# to append to the parsing function docstring below
+_doc_ts_formats = '\n'.join('\t- {}'.format(_) for _ in _TS_FORMATS)
+
+
+def _normalize_human_friendly_time(val, tz):
+    """Given one of :
+    - string (in one of the formats below)
+    - datetime (eg. datetime.now()), with or without tzinfo)
+    - timestamp (eg. time.time())
+    return a timestamp (seconds since jan 1 1970 UTC).
+
+    Non string/datetime values are returned unaltered.
+    Leading/trailing whitespace is stripped.
+    Supported formats:
+    {}
+    """
+    # {} is placeholder for formats; filled in after def...
+
+    zone = pytz.timezone(tz)  # tz as datetime.tzinfo object
+    epoch = pytz.UTC.localize(datetime(1970, 1, 1))
+    check = True
+
+    if isinstance(val, six.string_types):
+        # unix 'date' cmd format '%a %b %d %H:%M:%S %Z %Y' works but
+        # doesn't get TZ?
+
+        # Could cleanup input a bit? remove leading/trailing [ :,-]?
+        # Yes, leading/trailing whitespace to match pandas behavior...
+        # Actually, pandas doesn't ignore trailing space, it assumes
+        # the *current* month/day if they're missing and there's
+        # trailing space, or the month is a single, non zero-padded digit.?!
+        val = val.strip()
+
+        for fmt in _TS_FORMATS:
+            try:
+                ts = datetime.strptime(val, fmt)
+                break
+            except ValueError:
+                pass
+
+        try:
+            if isinstance(ts, datetime):
+                val = ts
+                check = False
+            else:
+                # what else could the type be here?
+                raise TypeError('expected datetime,'
+                                ' got {:r}'.format(ts))
+
+        except NameError:
+            raise ValueError('failed to parse time: ' + repr(val))
+
+    if check and not isinstance(val, datetime):
+        return val
+
+    if val.tzinfo is None:
+        # is_dst=None raises NonExistent and Ambiguous TimeErrors
+        # when appropriate, same as pandas
+        val = zone.localize(val, is_dst=None)
+
+    return (val - epoch).total_seconds()
+
+
+# fill in the placeholder we left in the previous docstring
+_normalize_human_friendly_time.__doc__ = (
+    _normalize_human_friendly_time.__doc__.format(_doc_ts_formats)
+)
